src/fanout/requirements.txt
Flask
Flask-Sock
redis
nats-py